Fix auto-layout, images bug, other issues.
The table wasn't using auto-layout so you could not scroll to the lower
rows of the table - it was off the screen.  Now fixed, should also mean
it works better on different sized screens.

Fix bug introduced with Swift 3.1 upgrade where the URLDataSessionTask
in the UIImage extension was not resumed after creation and thus did
nothing.

Fix code that altered the UI from the background thread - this will
crash the app and make it generally unstable - but does not work from
the autolayout engine.